--- src/window/bound_numeric_input.h
@@ -1,6 +1,0 @@
-#ifndef WINDOW_BOUND_NUMERIC_INPUT_H
-#define WINDOW_BOUND_NUMERIC_INPUT_H
-
-void window_bound_numeric_input_show(int x, int y, int max_digits, int min_value, int max_value, void (*callback)(int));
-
-#endif // WINDOW_BOUND_NUMERIC_INPUT_H
