+++ src/window/numeric_input.c
@@ -12,8 +12,11 @@
 #include "input/keyboard.h"
 #include "sound/effect.h"
 
+#include <limits.h>
+
 static void button_number(int number, int param2);
 static void button_accept(int param1, int param2);
+static void button_negative(int param1, int param2);
 static void button_delete(int param1, int param2);
 static void button_cancel(int param1, int param2);
 
@@ -32,15 +35,17 @@ static generic_button buttons[] = {
     {51, 111, 25, 25, button_number, button_none, 8, 0},
     {81, 111, 25, 25, button_number, button_none, 9, 0},
     {21, 141, 25, 25, button_number, button_none, 0, 0},
-    {51, 141, 55, 25, button_accept, button_none, 1, 0},
-    {21, 171, 25, 25, button_delete, button_none, 1, 0},
-    {51, 171, 55, 25, button_cancel, button_none, 1, 0}
+    {51, 141, 25, 25, button_negative, button_none, 0, 0},
+    {51, 171, 55, 25, button_accept, button_none, 1, 0},
+    {21, 201, 25, 25, button_delete, button_none, 1, 0},
+    {51, 201, 55, 25, button_cancel, button_none, 1, 0}
 };
 
 static struct {
     int x;
     int y;
     int max_digits;
+    int min_value;
     int max_value;
     void (*callback)(int);
 
@@ -49,12 +54,18 @@ static struct {
     int focus_button_id;
 } data;
 
-static void init(int x, int y, int max_digits, int max_value, void (*callback)(int))
+static void init(int x, int y, int max_digits, int min_value, int max_value, void (*callback)(int))
 {
     data.x = x;
     data.y = y;
     data.max_digits = max_digits;
-    data.max_value = max_value;
+    if (!min_value && !max_value) {
+        data.min_value = INT_MIN;
+        data.max_value = INT_MAX;
+    } else {
+        data.min_value = min_value;
+        data.max_value = max_value;
+    }
     data.callback = callback;
     data.num_digits = 0;
     data.value = 0;
@@ -80,7 +91,7 @@ static void draw_number_button(int x, int y, int number, int is_selected)
 
 static void draw_foreground(void)
 {
-    outer_panel_draw(data.x, data.y, 8, 14);
+    outer_panel_draw(data.x, data.y, 8, 15);
 
     graphics_fill_rect(data.x + 16, data.y + 16, 96, 30, COLOR_BLACK);
     if (data.num_digits > 0) {
@@ -98,14 +109,19 @@ static void draw_foreground(void)
     draw_number_button(data.x + 81, data.y + 111, 9, data.focus_button_id == 9);
     draw_number_button(data.x + 21, data.y + 141, 0, data.focus_button_id == 10);
 
-    graphics_draw_rect(data.x + 51, data.y + 141, 55, 25, data.focus_button_id == 11 ? COLOR_FONT_BLUE : COLOR_BLACK);
-    lang_text_draw_centered_colored(44, 16, data.x + 51, data.y + 147, 55, FONT_NORMAL_PLAIN,
+    if (data.min_value < 0) {
+        graphics_draw_rect(data.x + 51, data.y + 141, 25, 25, data.focus_button_id == 11 ? COLOR_FONT_BLUE : COLOR_BLACK);
+        text_draw_centered(string_from_ascii("-"), data.x + 51, data.y + 147, 25, FONT_NORMAL_PLAIN, COLOR_BLACK);
+    }
+    
+    graphics_draw_rect(data.x + 51, data.y + 171, 55, 25, data.focus_button_id == 12 ? COLOR_FONT_BLUE : COLOR_BLACK);
+    lang_text_draw_centered_colored(44, 16, data.x + 51, data.y + 177, 55, FONT_NORMAL_PLAIN,
             data.focus_button_id == 11 ? COLOR_FONT_BLUE : COLOR_BLACK);
 
-    graphics_draw_rect(data.x + 21, data.y + 171, 25, 25, data.focus_button_id == 12 ? COLOR_FONT_BLUE : COLOR_BLACK);
-    text_draw_centered(string_from_ascii("X"), data.x + 21, data.y + 177, 25, FONT_NORMAL_PLAIN, COLOR_RED);
-    graphics_draw_rect(data.x + 51, data.y + 171, 55, 25, data.focus_button_id == 13 ? COLOR_FONT_BLUE : COLOR_BLACK);
-    lang_text_draw_centered_colored(44, 17, data.x + 51, data.y + 177, 55, FONT_NORMAL_PLAIN,
+    graphics_draw_rect(data.x + 21, data.y + 201, 25, 25, data.focus_button_id == 12 ? COLOR_FONT_BLUE : COLOR_BLACK);
+    text_draw_centered(string_from_ascii("X"), data.x + 21, data.y + 207, 25, FONT_NORMAL_PLAIN, COLOR_RED);
+    graphics_draw_rect(data.x + 51, data.y + 201, 55, 25, data.focus_button_id == 13 ? COLOR_FONT_BLUE : COLOR_BLACK);
+    lang_text_draw_centered_colored(44, 17, data.x + 51, data.y + 207, 55, FONT_NORMAL_PLAIN,
             data.focus_button_id == 13 ? COLOR_FONT_BLUE : COLOR_BLACK);
 }
 
@@ -135,6 +151,11 @@ static void button_accept(int param1, int param2)
     input_accept();
 }
 
+static void button_negative(int param1, int param2)
+{
+    data.value = data.value * -1;
+}
+
 static void button_delete(int param1, int param2)
 {
     input_delete();
@@ -160,6 +181,9 @@ static void input_accept(void)
     if (data.value > data.max_value) {
         data.value = data.max_value;
     }
+    if (data.value < data.min_value) {
+        data.value = data.min_value;
+    }
     data.callback(data.value);
 }
 
@@ -174,12 +198,17 @@ static void input_delete(void)
 
 void window_numeric_input_show(int x, int y, int max_digits, int max_value, void (*callback)(int))
 {
+    window_numeric_input_bound_show(x, y, max_digits, 0, max_value, callback);
+}
+
+void window_numeric_input_bound_show(int x, int y, int max_digits, int min_value, int max_value, void (*callback)(int))
+{
     window_type window = {
         WINDOW_NUMERIC_INPUT,
         window_draw_underlying_window,
         draw_foreground,
         handle_input,
     };
-    init(x, y, max_digits, max_value, callback);
+    init(x, y, max_digits, min_value, max_value, callback);
     window_show(&window);
 }
