+++ src/window/numeric_input.h
@@ -3,6 +3,8 @@
 
 void window_numeric_input_show(int x, int y, int max_digits, int max_value, void (*callback)(int));
 
+void window_numeric_input_bound_show(int x, int y, int max_digits, int min_value, int max_value, void (*callback)(int));
+
 void window_numeric_input_accept(void);
 
 #endif // WINDOW_NUMERIC_INPUT_H
